% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze}
\alias{analyze}
\title{Loop the whole thing}
\usage{
analyze(df, col, x, y, p_value, id_col, filter_str = NULL, display_all = FALSE)
}
\arguments{
\item{df}{Dataframe containing biological data in long format}

\item{col}{Column of observations being tested}

\item{x}{Grouping column}

\item{y}{Numeric column used for kruskal-wallis and wilcoxon tests}

\item{p_value}{Numeric value between 0 and 1 determining significance}

\item{id_col}{Column with unique ids}

\item{filter_str}{Optional string to filter dataframe by -
col will be filtered by it}

\item{display_all}{TRUE/FALSE Default is FALSE. When FALSE, only the significant
results will be displayed}
}
\value{
Returns named list of results
}
\description{
Loop the whole thing
}
