% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{check_groups}
\alias{check_groups}
\title{Check group size}
\usage{
check_groups(df, group_col, id_col)
}
\arguments{
\item{df}{Dataframe containing groups being tested}

\item{group_col}{Column in data containing groups}

\item{id_col}{Column in data containing unique IDs}
}
\value{
Cleaned dataframe with groups removed that don't meet
the criteria of group size > 1.
}
\description{
In order for the other functions to be accurate,
the size of each group needs to be greater than 1.
}
