% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryOfPage.R
\name{countryOfPage}
\alias{countryOfPage}
\title{A countryOfPage Function}
\usage{
countryOfPage(idPage)
}
\arguments{
\item{idPage}{Facebook page id}
}
\description{
This function return most active country for page.
}
\examples{
topCountry <- countryOfPage("stunome")


}
\keyword{printLog}
