% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sheets.R
\name{create_attendance_sheet}
\alias{create_attendance_sheet}
\title{Create skeleton of Google Sheet for Attendance}
\usage{
create_attendance_sheet(code, schedule)
}
\arguments{
\item{assessment}{A tibble with columns "Assessment" that depicts the name of the assessment, "Weightage" depicts the weightage of the assessment and "Total Marks" is the total marks alloted for the assessment.}
}
\description{
Function to create a google sheets for storing information on attendance of students.
}
