% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sheets.R
\name{create_authorization_sheet}
\alias{create_authorization_sheet}
\title{Create skeleton of Google Sheet for Authorisation for web application access}
\usage{
create_authorization_sheet(code, assessment)
}
\description{
Functions to create a google sheets for storing information on the faculty members authorized to access the web application.
}
