% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sheets.R
\name{create_grade_sheet}
\alias{create_grade_sheet}
\title{Create skeleton of Google Sheet for Assessment Grades}
\usage{
create_grade_sheet(code, assessment)
}
\arguments{
\item{schedule}{A tibble with columns "Class","StartDate","EndDate","By" where "Class" depicts the type of class ie.Lecture A, Lecture B, Tutorial A, Tutorial B etc. "StartDate" and "EndDate" depicts the date of start and end of the class in the semester. "By" depicts the number of classes in a week.}
}
\description{
Functions to create a google sheets for storing information on Grade.
}
