% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sheets.R
\name{create_student_sheet}
\alias{create_student_sheet}
\title{Create skeleton of Google Sheets}
\usage{
create_student_sheet()
}
\description{
Functions to create a google sheets for storing information on Attendance, Grade and Authorisation list.
}
