% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{create_time_data}
\alias{create_time_data}
\title{Create one dataframe with all timepoints}
\usage{
create_time_data(...)
}
\arguments{
\item{...}{A list of files to read in. These should already be clean and all files need to include the same columns in the same order
It's important as well that the files are inserted in order that they were created, i.e. the earliest timepoint should be first, second
timepoint should be second, etc.}
}
\value{
A single dataframe that lists all samples for each timepoint.
}
\description{
Create one dataframe with all timepoints
}
