% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptionOfPage.R
\name{descriptionOfPage}
\alias{descriptionOfPage}
\title{A descriptionOfPage Function}
\usage{
descriptionOfPage(idPage)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)


}
\keyword{printLog}
