% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbOAuthSugar.R
\name{fbOAuthSugar}
\alias{fbOAuthSugar}
\title{A fbOAuth Function}
\usage{
fbOAuthSugar(app_id, app_secret, extended_permissions = FALSE,
  legacy_permissions = FALSE, scope = NULL)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}

