% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComments.R
\name{get_comments}
\alias{get_comments}
\title{A get_comments Function}
\usage{
get_comments(toAnalyze, maxComments = 1)
}
\arguments{
\item{toAnalyze}{vector with posts ids.}

\item{maxComments}{max count of likes under posts}
}
\description{
This function rerturn data frame with commenters and comments for vector with posts Ids.
}
\examples{
toAnalyze <- "90002267161_10155520966952162"
dfCommenters <- get_comments(toAnalyze, 1000)

}
\keyword{printLog}
