% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_engagers.R
\name{get_engagers}
\alias{get_engagers}
\title{A get_engagers Function}
\usage{
get_engagers(toAnalyze, maxLikes = 1)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}

