% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_likers.R
\name{get_likers}
\alias{get_likers}
\title{A get_likers Function}
\usage{
get_likers(toAnalyze, write = FALSE, maxLikes = 1)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}

