% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{get_percs}
\alias{get_percs}
\title{Calcuate glycan component percentages}
\usage{
get_percs(data, glycan, id_col)
}
\arguments{
\item{glycan}{The specific binary glycan component column}

\item{id_col}{The column that holds the unique ids of samples

The dataframe fed into this function must include the columns `Area`,
`adj_area`, `total_area`, `adj_tot_area`.}

\item{df}{The dataframe}
}
\value{
A dataframe with a new column for the abundance of glycan component
}
\description{
Calcuate glycan component percentages
}
