% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{get_tots}
\alias{get_tots}
\title{Calculate total relative abundance for glycan component}
\usage{
get_tots(df, glycan, unique_id, species)
}
\arguments{
\item{df}{Dataframe including columns `adj_area` and `adj_tot_area`}

\item{glycan}{Glycan component that total abundance is being calculated}

\item{unique_id}{Column containing unique id for each sample}

\item{species}{Column containing the different glycan species}
}
\value{
The dataframe used in the function with an additional column with
the total abundance of glycan components that have both 1 or 2 component
options.
}
\description{
Calculate total relative abundance for glycan component
}
