% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttendingUsers.R
\name{getAttendingUsers}
\alias{getAttendingUsers}
\title{A getAttendingUsers Function}
\usage{
getAttendingUsers(id)
}
\arguments{
\item{id}{event id from Facebook}
}
\description{
This function return list of people who are attendinge event.
}
\examples{
users <- getAttendingUsers("1972457919653234")

}
\keyword{printLog}
