% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFacebookPageEvents.R
\name{getFacebookPageEvents}
\alias{getFacebookPageEvents}
\title{A getFacebookPageEvents Function}
\usage{
getFacebookPageEvents(forUsername)
}
\arguments{
\item{forUsername}{Facebook id or Facebook username}
}
\description{
This function return last 25 events for Facebook Page
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}
