% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInerestedUsers.R
\name{getInerestedUsers}
\alias{getInerestedUsers}
\title{A getAttendingUsers Function}
\usage{
getInerestedUsers(id)
}
\arguments{
\item{id}{event id from Facebook}
}
\description{
This function return list of people who are interested in event.
}
\examples{
users <- getAttendingUsers("1972457919653234")

}
\keyword{printLog}
