% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPostsNew.R
\name{getPageNew}
\alias{getPageNew}
\title{A getPageNew Function}
\usage{
getPageNew(idPage, posts = 25)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}

