% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPostsNew.R
\name{getPagePosts}
\alias{getPagePosts}
\title{A getPagePosts Function}
\usage{
getPagePosts(idPage, interations = 10)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}
