% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPagePostsByDate.R
\name{getPagePostsByDate}
\alias{getPagePostsByDate}
\title{A getPagePostsByDate Function}
\usage{
getPagePostsByDate(idPage, fromDate, toDate, interations = 10000)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}
