% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPagePostsInteractionsByDate.R
\name{getPagePostsInteractionsByDate}
\alias{getPagePostsInteractionsByDate}
\title{A getPagePostsInteractionsByDate Function}
\usage{
getPagePostsInteractionsByDate(idPage, fromDate, toDate, interations = 10000)
}
\arguments{
\item{idPage}{page id}

\item{fromDate}{first day of period}

\item{toDate}{last day of period}
}
\description{
This function return besic info about pages posts by date. This function is just part of getPageNewByDate function
}
\examples{
posts <- getPagePostsInteractionsByDate("stunome")


}
\keyword{printLog}
