% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPagePostsInteractionsDetail.R
\name{getPagePostsInteractionsDetail}
\alias{getPagePostsInteractionsDetail}
\title{A getPagePostsInteractionsDetail Function}
\usage{
getPagePostsInteractionsDetail(id)
}
\arguments{
\item{id}{Facebook post id}
}
\description{
This function return number od reactions on specifick post.
}
\examples{
reactions <- getPagePostsInteractionsDetail()


}
\keyword{printLog}
