% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPostAlternative.R
\name{getPostAlternative}
\alias{getPostAlternative}
\title{A getPostAlternative Function}
\usage{
getPostAlternative(idPage, timestamp = "1420070400")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}
