% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPostsByDate.R
\name{getPostsByDate}
\alias{getPostsByDate}
\title{A getPageNewByDate Function}
\usage{
getPostsByDate(idPage, fromDate, toDate)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function is depricated
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}
