% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReactions.R
\name{getReactions}
\alias{getReactions}
\title{A getReactions Function}
\usage{
getReactions(id, type = "all")
}
\arguments{
\item{id}{Facebook pages id.}
}
\description{
This function return data frame with list of engagers under post.
}
\examples{
postEngagers <- getReactions("stunome")


}
\keyword{printLog}
