% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{group_test}
\alias{group_test}
\title{Between groups testing}
\usage{
group_test(df, x, y, id_col, filter_col = NULL, filter_str = NULL)
}
\arguments{
\item{df}{Dataframe containing groups being tested}

\item{x}{Independent variable for kruskal-wallis test}

\item{y}{Dependent variable for kruskal-wallis test}

\item{filter_col}{Column filtered on}

\item{filter_str}{Filtering condition}
}
\value{
List containing kruskal-wallis results and visualization
}
\description{
This function takes a dataframe and uses a
kruskal-wallis test to determine if groups
come from the same distribution or not.
}
