% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_auth.R
\name{initiate_auth}
\alias{initiate_auth}
\title{Initiate the authorization}
\usage{
initiate_auth(setup_info)
}
\arguments{
\item{setup_info}{app set up information}
}
\description{
Function to set up googlesheets authorisation
}
