% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Shiny Unit Grade and Attendance Reviewer}
\usage{
launch_app()
}
\description{
Launches the shiny web application that allows students to see their grade and attendance of a unit.
}
