% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOfPages.R
\name{listOfPages}
\alias{listOfPages}
\title{A listOfPages Function}
\usage{
listOfPages(idPage)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function return list of pages liked by page.
}
\examples{

liked <- listOfPages("stunome")


}
\keyword{printLog}
