% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkOfPages.R
\name{networkOfPages}
\alias{networkOfPages}
\title{A networkOfPages Function}
\usage{
networkOfPages(toAnalyze, deep = 3, saveToTable = FALSE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)


}
\keyword{printLog}

