% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapOutput.R
\name{overlapOutput}
\alias{overlapOutput}
\title{A overlapOutput Function}
\usage{
overlapOutput(toAnalyze, mainPageList, analyzedFrom, analyzedTo, backgroundFrom,
  backgroundTo, comments = "F")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function return csv files with Smart Overlap
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}
