% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{pair_test}
\alias{pair_test}
\title{Pairwise Comparisons}
\usage{
pair_test(
  df,
  values,
  groups,
  p_value,
  id_col,
  filter_str = NULL,
  filter_col = NULL
)
}
\arguments{
\item{df}{Dataframe doing comparisons on}

\item{values}{Numeric column of df}

\item{groups}{Factor or character column of df}

\item{p_value}{Value between 0 and 1 indicating level of significance}

\item{id_col}{Column of unique IDs used to check group sizes}

\item{filter_str}{Optional string supplied if filtering within function}

\item{filter_col}{Optional column supplied if filtering within function.
Must be present if filter_str is supplied}
}
\value{
Returns a matrix object
}
\description{
Do all the pairwise comparisons in one swoop
}
