% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_overall}
\alias{plot_overall}
\title{Plot barcharts for overall comparison}
\usage{
plot_overall(df, x, y, group)
}
\arguments{
\item{df}{Dataframe}

\item{x}{It should be a column of categorical data}

\item{y}{Column for the y axis - numeric}

\item{group}{Group for color - categorical}
}
\value{
list of plots
}
\description{
Plot barcharts for overall comparison
}
