% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{plot_time_data}
\alias{plot_time_data}
\title{Visualize trends over time}
\usage{
plot_time_data(
  df,
  x,
  y,
  id_col,
  color_col = NULL,
  filter_str = NULL,
  filter_col = NULL
)
}
\arguments{
\item{df}{A dataframe with multiple timepoints}

\item{x}{The column that will provide the x axis with data. In the case
of this function, it should be some measure of time.}

\item{y}{Column that will provide the y axis with data.}

\item{id_col}{Column that indicates the grouping factor for connecting
points}

\item{color_col}{Optional column to determine how the plot should be colored}

\item{filter_str}{Optional argument that allows the dataframe to be filtered
within the function - provide the desired filtering string here}

\item{filter_col}{Column being filtered within the dataframe. If filter_str
is provided, this also must be present}
}
\value{
A single dataframe that lists all samples for each timepoint.
}
\description{
Visualize trends over time
}
