% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printLog.R
\name{printLog}
\alias{printLog}
\title{A printLog Function}
\usage{
printLog(...)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)
}
\keyword{printLog}

