% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_info}
\alias{setup_info}
\title{Create app set up information}
\usage{
setup_info(unitinfo, maintainer, client_id, client_secret)
}
\arguments{
\item{unitinfo}{unit information}

\item{maintainer}{email id of the maintainer of the application}

\item{client_id}{Google cloud platform client id}

\item{client_secret}{Google cloud platform client secret}
}
\value{
a special class, app_setup with the set up information
}
\description{
Function to recieve unit information
}
