% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugaR.R
\name{sugaR}
\alias{sugaR}
\title{sugaR}
\usage{
sugaR(name, auth_type, endpoint = NULL, cachePath = "~\\\\")
}
\arguments{
\item{name}{name of the service}

\item{auth_type}{authentication type: "none", "basic", ...}

\item{endpoint}{endpoint in case auth_type is oauth1 or oauth2}

\item{cachePath}{path of the cache folder}
}
\description{
sugaR
}
\examples{
\dontrun{
tmp  <- sugaR()
}
}
\keyword{sugaR}

