\name{u}
\Rdversion{1.1}
\alias{u}
\title{Convert time of day, denoted as HH:MM, to double as HH+MM/60}
\description{
This function helps with the entry of data for the interpretation of blood glucose levels.
}
\usage{
u(X)
}
\arguments{
\item{X}{String to convert.}
}
\seealso{
\code{\link{h}}.
}
\examples{
u("8:47")
}
\author{Steffen Moeller \email{steffen_moeller@gmx.de}}

