% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sheets.R
\name{create_attendance_sheet}
\alias{create_attendance_sheet}
\title{Create skeleton of Google Sheet for Attendance}
\usage{
create_attendance_sheet(unitinfo, schedule)
}
\arguments{
\item{unitinfo}{unit information}

\item{schedule}{A tibble with columns "Class","StartDate" and "EndDate" where "Class" depicts the type of class ie.Lecture A, Lecture B, Tutorial A, Tutorial B etc. "StartDate" and "EndDate" depicts the date of start and end of the class in the semester.}
}
\description{
Function to create a google sheets for storing information on attendance of students.
}
