% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sheets.R
\name{create_grade_sheet}
\alias{create_grade_sheet}
\title{Create skeleton of Google Sheet for Assessment Grades}
\usage{
create_grade_sheet(unitinfo, assessment)
}
\arguments{
\item{unitinfo}{unit information
@param assessment A tibble with columns "Assessment" that depicts the name of the assessment, "Weightage" depicts the weightage of the assessment and "Total Marks" is the total marks alloted for the assessment.}
}
\description{
Function to create a google sheets for storing information on Grade.
}
