% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sheets.R
\name{create_student_sheet}
\alias{create_student_sheet}
\title{Create skeleton of Google Sheet for student}
\usage{
create_student_sheet()
}
\description{
Function to create a google sheets for storing information on Students.
}
