% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComments.R
\name{get_comments}
\alias{get_comments}
\title{A get_comments Function}
\usage{
get_comments(toAnalyze, maxComments = 1)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}

