% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_likers.R
\name{get_likers}
\alias{get_likers}
\title{A get_likers Function}
\usage{
get_likers(toAnalyze, maxLikes = 1)
}
\arguments{
\item{toAnalyze}{vector with posts ids.}

\item{maxLikes}{max count of likes under posts}
}
\description{
This function rerturn data frame with likers for specic vector with posts Ids.
}
\examples{
toAnalyze <- "90002267161_10155520966952162"
dfLikers <- get_likers(toAnalyze, 1000)

}
\keyword{printLog}
