% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFacebookPageEvents.R
\name{getFacebookPageEvents}
\alias{getFacebookPageEvents}
\title{A getFacebookPageEvents Function}
\usage{
getFacebookPageEvents(forUsername, access_token)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}

