% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFacebookPageSearch.R
\name{getFacebookPageSearch}
\alias{getFacebookPageSearch}
\title{A getFacebookPageSearch Function}
\usage{
getFacebookPageSearch(query)
}
\arguments{
\item{query}{search query}
}
\description{
This function return first 25 results for Facebook Page search
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}
