% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPagesDetail.R
\name{getPagesDetail}
\alias{getPagesDetail}
\title{A getPagesDetail Function}
\usage{
getPagesDetail(toAnalyze)
}
\arguments{
\item{toAnalyze}{vector with id of Facebook pages}
}
\description{
This function return basic info about pages. Included id, name, about, likes and talling talking_about_count
}
\examples{
toAnalyze <- c("stunome")
dfPages <- getPagesDetail(toAnalyze)


}
\keyword{printLog}
