% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReactions.R
\name{getReactions}
\alias{getReactions}
\title{A getReactions Function}
\usage{
getReactions(id, type = "all")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}

