% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_auth.R
\name{initiate_auth}
\alias{initiate_auth}
\title{Initiate the authorization}
\usage{
initiate_auth(email)
}
\arguments{
\item{email}{email id of the google sheet user}
}
\description{
Function to set up googlesheets authorisation
}
