% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkOfPages.R
\name{networkOfPages}
\alias{networkOfPages}
\title{A networkOfPages Function}
\usage{
networkOfPages(toAnalyze, deep = 2)
}
\arguments{
\item{toAnalyze}{vector with Facebook pages ids.}

\item{deep}{how many steps you go deeper, default is 2}
}
\description{
This function download data frame with pages liked by page network. Results are stored in succesNetwork.txt data table. 
Function return also data frame with same value.
}
\examples{
toAnalyze <- c("stunome")
network <- networkOfPages(toAnalyze, 1)


}
\keyword{printLog}
