% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_info}
\alias{setup_info}
\title{Create app set up information}
\usage{
setup_info(
  unit = unitinfo$unit,
  unit_code = unitinfo$unit_code,
  name = unitinfo$name,
  maintainer,
  client_id,
  client_secret
)
}
\arguments{
\item{maintainer}{email id of the maintainer of the application}

\item{client_id}{Google cloud platform client id}

\item{client_secret}{Google cloud platform client secret}
}
\value{
a special class, app setup with the set up as a list
}
\description{
Function to recieve unit information
}
\examples{
setup_info(unit,maintainer=c("Aarathy Babu"="abab0012@student.monash.edu"))

}
