% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_info.R
\name{unit_info}
\alias{unit_info}
\title{Create unit information}
\usage{
unit_info(unit, name, semester, year, lecturer)
}
\arguments{
\item{unit}{the unit code}

\item{name}{string that denotes the name of the unit}

\item{semester}{Integer that denotes teaching period}

\item{year}{Year}

\item{lecturer}{character vector with name and email id of lecturers}
}
\value{
a special class, monash unit with the unit information as a list
}
\description{
Function to recieve unit information
}
